Implemented the SHAP analyser.
Implemented the SHAP analyser which computes the SHAP values and displays the result as a bar plot in a seperate window.

NOTE: Currently, only the first target class' SHAP values are being displayed on the bar plot. Need to figure out how to display an overview of the entire model/data.